samples/pyproject.toml
[tool.poetry]
name = "dfclean-demo"
version = "0.1.0"
description = "Demo project loading a CSV and running dfclean pipeline"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{ include = "dfclean_sample" }]

[tool.poetry.dependencies]
python = ">=3.11,<4.0"
pandas = ">=2.0"
dfclean = { git = "https://github.com/arappaport/dfclean.git", branch = "main" }
pyyaml = "^6.0.3"

[tool.poetry.group.dev.dependencies]
pytest = ">=8.0"
ruff = ">=0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
